Fixed: file structure, all const wete moved to constants.js
