+++ src/components/popupWithForm.js
@@ -4,7 +4,7 @@ export default class PopupWithForm extends Popup {
   constructor({popupSelector, formSubmit}) {
     super(popupSelector);
     this._formSubmit = formSubmit;
-    
+
     this._popupForm = this._popup.querySelector('.popup__form');
     this._popupInput = this._popupForm.querySelectorAll('.popup__input');
   }
@@ -26,8 +26,8 @@ export default class PopupWithForm extends Popup {
   }
 
   closePopup() {
-  super.closePopup();
-  this._popupForm.reset();
+    super.closePopup();
+    this._popupForm.reset();
   }
 
 }
