+++ src/components/validate.js
@@ -9,71 +9,70 @@ export default class FormValidator {
 
 // Валидация формы "Редактировать профиль"
 // Ищем элемент в форме и добавляем ему класс с ошибкой
-_showInputError(inputElement, errorMessage) {
-  const errorElement = this._formElement.querySelector(`.${inputElement.id}-error`);
-  inputElement.classList.add(this._config.inputErrorClass);
-  errorElement.textContent = errorMessage;
-  errorElement.classList.add(this._config.errorClass);
-};
+  _showInputError(inputElement, errorMessage) {
+    const errorElement = this._formElement.querySelector(`.${inputElement.id}-error`);
+    inputElement.classList.add(this._config.inputErrorClass);
+    errorElement.textContent = errorMessage;
+    errorElement.classList.add(this._config.errorClass);
+  };
 
 // Ищем элемент в форме и убираем класс с ошибкой
-_hideInputError (inputElement) {
-  const errorElement = this._formElement.querySelector(`.${inputElement.id}-error`);
-  inputElement.classList.remove(this._config.inputErrorClass);
-  errorElement.classList.remove(this._config.errorClass);
-  errorElement.textContent = '';
-};
+  _hideInputError(inputElement) {
+    const errorElement = this._formElement.querySelector(`.${inputElement.id}-error`);
+    inputElement.classList.remove(this._config.inputErrorClass);
+    errorElement.classList.remove(this._config.errorClass);
+    errorElement.textContent = '';
+  };
 
 //Проверяем вводимый элемент на валидарность и показываем ошибку, если false
-_checkInputValidity(inputElement) {
-  if (inputElement.validity.patternMismatch) {
-    inputElement.setCustomValidity(inputElement.dataset.errorMessage);
-  } else {
-    inputElement.setCustomValidity("");
-  }
-  if (!inputElement.validity.valid) {
-    this._showInputError(inputElement, inputElement.validationMessage);
-  } else {
-    this._hideInputError(inputElement);
-  }
-};
+  _checkInputValidity(inputElement) {
+    if (inputElement.validity.patternMismatch) {
+      inputElement.setCustomValidity(inputElement.dataset.errorMessage);
+    } else {
+      inputElement.setCustomValidity("");
+    }
+    if (!inputElement.validity.valid) {
+      this._showInputError(inputElement, inputElement.validationMessage);
+    } else {
+      this._hideInputError(inputElement);
+    }
+  };
 
 // Ищем все поля input и кнопку submit
-_setEventListeners()
-{
+  _setEventListeners() {
 
-  // Делаем кнопку неактивной перед первым вводом
-  this.toggleButtonState();
+    // Делаем кнопку неактивной перед первым вводом
+    this.toggleButtonState();
 
-  // Проверяем все вводимые элементы на ошибки и регулирует активность кнопки
-  this._inputList.forEach((inputElement) => {
+    // Проверяем все вводимые элементы на ошибки и регулирует активность кнопки
+    this._inputList.forEach((inputElement) => {
       inputElement.addEventListener('input', () => {
         this._checkInputValidity(inputElement);
         this.toggleButtonState();
       });
     });
-};
+  };
 
 // Ищем все формы и отменяем стандартные визуализации ошибки браузера
-enableValidation() {
-  this._setEventListeners();
-}
+  enableValidation() {
+    this._setEventListeners();
+  }
 
 //Проверяем вводимый элемент на валидарность посредством input и возвращаем результат
-_hasInvalidInput() {
-  return this._inputList.some((inputElement) => {
-    return !inputElement.validity.valid;
-  })
-};
+  _hasInvalidInput() {
+    return this._inputList.some((inputElement) => {
+      return !inputElement.validity.valid;
+    })
+  };
 
 // Присваиваем все элементы на валидарность, если хоть один невыполненн, то делаем кнопку неактивной
-toggleButtonState() {
-  if (this._hasInvalidInput()) {
-    this._buttonElement.classList.add(this._config.inactiveButtonClass);
-    this._buttonElement.setAttribute('disabled', true);
-  } else {
-    this._buttonElement.classList.remove(this._config.inactiveButtonClass);
-    this._buttonElement.removeAttribute('disabled');
-  }
-};
+  toggleButtonState() {
+    if (this._hasInvalidInput()) {
+      this._buttonElement.classList.add(this._config.inactiveButtonClass);
+      this._buttonElement.setAttribute('disabled', true);
+    } else {
+      this._buttonElement.classList.remove(this._config.inactiveButtonClass);
+      this._buttonElement.removeAttribute('disabled');
+    }
+  };
 }
